+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/video/MediaCodecVideoRenderer.java
@@ -1808,14 +1808,14 @@ public class MediaCodecVideoRenderer extends MediaCodecRenderer
       // Make sure to decode and render the last frame.
       return false;
     }
-    boolean shouldSkipDecoderInputBuffer = isBufferBeforeStartTime(buffer);
+    boolean isBufferBeforeStartTime = isBufferBeforeStartTime(buffer);
     boolean shouldDropDecoderInputBuffer = false;
     if (videoFrameReleaseEarlyTimeForecaster != null) {
       long predictedEarlyUs = videoFrameReleaseEarlyTimeForecaster.predictEarlyUs(buffer.timeUs);
       shouldDropDecoderInputBuffer =
           predictedEarlyUs != C.TIME_UNSET && predictedEarlyUs < minEarlyUsToDropDecoderInput;
     }
-    if (!shouldSkipDecoderInputBuffer && !shouldDropDecoderInputBuffer) {
+    if (!isBufferBeforeStartTime && !shouldDropDecoderInputBuffer) {
       return false;
     }
     if (buffer.hasSupplementalData()) {
@@ -1829,7 +1829,7 @@ public class MediaCodecVideoRenderer extends MediaCodecRenderer
         && checkNotNull(getCodecInfo()).mimeType.equals(MimeTypes.VIDEO_AV1)
         && buffer.data != null) {
       boolean skipFrameHeaders =
-          shouldSkipDecoderInputBuffer
+          isBufferBeforeStartTime
               || consecutiveDroppedInputBufferCount
                   <= MAX_CONSECUTIVE_DROPPED_INPUT_BUFFERS_COUNT_TO_DISCARD_HEADER;
       ByteBuffer readOnlySample = buffer.data.asReadOnlyBuffer();
@@ -1852,9 +1852,7 @@ public class MediaCodecVideoRenderer extends MediaCodecRenderer
       }
     }
     if (bufferDiscarded) {
-      if (shouldSkipDecoderInputBuffer) {
-        decoderCounters.skippedInputBufferCount += 1;
-      } else {
+      if (!isBufferBeforeStartTime) {
         consecutiveDroppedInputBufferCount += 1;
       }
       discardedDecoderInputBufferTimestamps.add(buffer.timeUs);
@@ -2295,20 +2293,27 @@ public class MediaCodecVideoRenderer extends MediaCodecRenderer
       Log.d(DEBUG_LOG_TAG, "video, discard input to keyframe, count=" + droppedSourceBufferCount);
     }
     int droppedInQueue = 0;
+    int skippedInQueue = 0;
+    boolean isJoining = videoFrameReleaseControl.isJoining();
     for (long timeUs : discardedDecoderInputBufferTimestamps) {
-      if (timeUs >= getLastResetPositionUs()) {
+      if (timeUs >= getLastResetPositionUs() && !isJoining) {
         droppedInQueue++;
+      } else {
+        skippedInQueue++;
       }
     }
+    discardedDecoderInputBufferTimestamps.clear();
     if (treatDroppedBuffersAsSkipped) {
       decoderCounters.skippedInputBufferCount += droppedSourceBufferCount;
       decoderCounters.skippedOutputBufferCount += buffersInCodecCount;
       decoderCounters.skippedInputBufferCount += droppedInQueue;
+      decoderCounters.skippedInputBufferCount += skippedInQueue;
     } else {
       decoderCounters.droppedToKeyframeCount++;
       updateDroppedBufferCounters(
           /* droppedInputBufferCount= */ droppedSourceBufferCount + droppedInQueue,
           /* droppedDecoderBufferCount= */ buffersInCodecCount);
+      decoderCounters.skippedInputBufferCount += skippedInQueue;
     }
     flushOrReinitializeCodec();
     if (videoSink != null) {
@@ -2347,16 +2352,21 @@ public class MediaCodecVideoRenderer extends MediaCodecRenderer
    */
   private void updateDiscardedBufferCountersWithInputBuffers(long bufferPresentationTimeUs) {
     int droppedInputBufferCount = 0;
+    int skippedInputBufferCount = 0;
     Long minDiscardedDecoderBufferTimeUs;
+    boolean isJoining = videoFrameReleaseControl.isJoining();
     while ((minDiscardedDecoderBufferTimeUs = discardedDecoderInputBufferTimestamps.peek()) != null
         && minDiscardedDecoderBufferTimeUs < bufferPresentationTimeUs) {
       discardedDecoderInputBufferTimestamps.poll();
       frameRateEstimator.onNextFrame(minDiscardedDecoderBufferTimeUs * 1000);
-      if (minDiscardedDecoderBufferTimeUs >= getLastResetPositionUs()) {
+      if (minDiscardedDecoderBufferTimeUs >= getLastResetPositionUs() && !isJoining) {
         droppedInputBufferCount++;
+      } else {
+        skippedInputBufferCount++;
       }
     }
     updateDroppedBufferCounters(droppedInputBufferCount, /* droppedDecoderBufferCount= */ 0);
+    decoderCounters.skippedInputBufferCount += skippedInputBufferCount;
   }
 
   /**