+++ RELEASENOTES.md
@@ -154,6 +154,8 @@
         ([#418](https://github.com/androidx/media/issues/418)).
     *   Add support for big endian 32-bit and 64-bit floating point PCM.
 *   Video:
+    *   Fix video joining logic to count dropped buffers during joining as
+        skipped instead of dropped.
     *   Fix immediate rendering decision when setting new surface to avoid frame
         drops on devices that don't support placeholder surfaces.
     *   Fix frame rate estimation in `MediaCodecVideoRenderer` to be more